els-ASAP auto commit file/tag eclipse-rse-3.2-1.el6.src.rpm

+++ src/src/org/eclipse/rse/internal/services/local/files/LocalFileService.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2006, 2009 IBM Corporation and others.
+ * Copyright (c) 2006, 2010 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -44,6 +44,8 @@
  * David McKnight   (IBM)        - [280899] RSE can't open files in some directory, which give the RSEG1067 error me
  * Martin Oberhuber (Wind River) - [285942] Throw exception when listing a non-folder
  * Martin Oberhuber (Wind River) - [286129][api] RemoteFileException(String) violates API contract
+ * David McKnight   (IBM)        - [299140] Local Readonly file can't be copied/pasted twice
+ * Martin Oberhuber (Wind River) - [314461] NPE deleting a folder w/o permission
  *******************************************************************************/
 
 package org.eclipse.rse.internal.services.local.files;
@@ -405,6 +407,13 @@ public class LocalFileService extends AbstractFileService implements ILocalServi
 
 			inputStream = new FileInputStream(file);
 			bufInputStream = new BufferedInputStream(inputStream);
+			
+			boolean wasReadonly = destinationFile.exists() && !destinationFile.canWrite();
+			if (wasReadonly){ // tempfile is readonly
+				// since we're replacing the tempfile that represents the real file, the readonly bit should be removed for the transfer
+				//destinationFile.setWritable(true);
+				setReadOnly(destinationFile.getParent(), destinationFile.getName(), false, monitor);
+			}
 			outputStream = new FileOutputStream(destinationFile);
 
 			if (isEncodingConversionRequired)
@@ -454,6 +463,9 @@ public class LocalFileService extends AbstractFileService implements ILocalServi
 					monitor.worked(bytesRead);
 					isCancelled = monitor.isCanceled();
 				}
+			}
+			if (wasReadonly){
+				destinationFile.setReadOnly();
 			}
 		}
 		catch (Exception e)
@@ -1090,16 +1102,20 @@ public class LocalFileService extends AbstractFileService implements ILocalServi
 	{
 		boolean result = true;
 		File[] files = folder.listFiles();
-		for (int i = 0; i < files.length && result; i++)
+		if (files!=null)
 		{
-			File file = files[i];
-			if (file.isDirectory())
+			//bug 314461: java.io.File returns null if folder has no permissions
+			for (int i = 0; i < files.length && result; i++)
 			{
-				result = deleteContents(file, monitor);
-			}
-			else
-			{
-				result = file.delete();
+				File file = files[i];
+				if (file.isDirectory())
+				{
+					result = deleteContents(file, monitor);
+				}
+				else
+				{
+					result = file.delete();
+				}
 			}
 		}
 		if (result)
